Create a character set
We need a set that can represent the entire Unicode character range
without being wasteful of memory, like Python's set type would do.  We
also need to be able to calculate the disjoint representation of a
bunch of character sets.  This provides an implementation of a set type
that uses ranges of characters (actually represented as integers), and
additionally provides a disjoint() class method to calculate the
disjoint representation.  The disjoint() method uses a priority queue
(implemented on top of Python's standard heapq package) for its
algorithm.

Full tests are included.  Note that the CharSet class could be further
optimized by overriding some of the default implementations of the
methods provided by the base class, collections.MutableSet; as CharSet
is used deep in the core of DFA construction, it may prove to be a hot
spot where such optimizations could be useful.