Create a state representation
The plexgen lexer generator uses a finite state automaton to perform the
actual lexing.  Of course, FSAs need states.  This class provides the
basic representation of an automaton state, including utility routines
to add transitions, reverse a state, and examine whether a state has only
Epsilon transitions coming in to it, not to mention iterating over
selected transitions into or out of a state.  Full tests are included.