Create state transitions
The plexgen lexer generator uses a finite state automaton to perform the
actual lexing.  Of course, FSAs need to be able to transition from one
state to another.  This commit provides a set of classes to represent the
state transitions, including the ability to merge similar transitions and
to simulate running the transition against an input.  Full tests are
included.